refactor(web): extract use-edit-session-state.ts from SpatialEditor
The open-dialog and pending-edit state (edgeLabelEditId,
groupLabelEditId, linkDialog, canvasPicker, facetPanelOpen + its
render-time clear, pendingCut + its invalidation effect,
selectedEdgeId, showResolvedComments) moves verbatim into its own
hook, called ahead of useWorkerScene: none of it is scene-derived,
and showResolvedComments is itself an input to the worker-scene call.
The render-time facetPanelOpen clear stays a render-time clear inside
the hook body, not an effect.

editor-state-surface.test.ts's sources glob gains
use-edit-session-state.ts.
